# backend/types/models/BaseObject.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * Object representing information that is reused often. This is not meant to be used in api, just as an helper to avoid duplication.
 */
export type BaseObject = {
    id: number;
    name: string;
}