# backend/types/models/Company.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { BaseObject } from './BaseObject';
import type { Project } from './Project';

export type Company = (BaseObject & Array<Project>);