# openapi-definition.yaml
...
openapi: 3.0.3
info:
  version: 0.0.1
  title: Tuntikikkare
  license:
    name: MIT
paths:
  /companies:
    get:
      summary: Get all companies
      operationId: getCompanies
      responses:
        '200': 
          description: List of companies added to the system
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Companies"
    post: 
      summary: Create a company
      operationId: addCompany
      requestBody:
        description: Company information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyAdd'
      responses:
        '201':
          description: Company added
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        default:  
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    BaseObject: 
      type: object
      description: Object representing information that is reused often. This is not meant to be used in api, just as an helper to avoid duplication.
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          minLength: 1
          maxLength: 255
    Companies:
      type: array
      items:
        $ref: '#/components/schemas/Company'
    Company:
      allOf:
        - $ref: '#/components/schemas/BaseObject'
        - type: array
          description: List of projects of this company
          items: 
            $ref: "#/components/schemas/Project"
    CompanyAdd:
      type: object
      description: Object representing a new company
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
    Project:
      allOf:
        - $ref: '#/components/schemas/BaseObject'
        - type: object
          description: Object representing a project
          required:
            - company_id
            - tasks
          properties:
            company_id:
              type: integer
              format: int32
            tasks:
              type: array
              items: 
                $ref: "#/components/schemas/Task"
...
    Task:
      allOf:
        - $ref: '#/components/schemas/BaseObject'
        - type: object
          description: Object representing a task. 
          required:
            - project_id
            - entries
          properties:
            project_id:
              type: integer
              format: int32
            entries:
              type: array
              items:  
                $ref: "#/components/schemas/Entry"
...
    Entry:
      type: object
      description: Object that represents an entry that an employee works on in a project
      required:
        - id
        - employee_id
        - task_id
        - date
        - hours
      properties:
        id:
          type: integer
          format: int32
        employee_id:
          type: integer
          format: int32
        task_id:
          type: integer
          format: int32
        date: 
          type: string
          format: date
          description: Date when the work has been done
        hours: 
          type: number
          format: double
          description: How many hours of work has been done
        comment:
          type: string
          description: Optional comment about the entry
          minLength: 1
          maxLength: 255
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
          description: Error message
          minLength: 1
          maxLength: 255